vk2telegram/requirements.txt
python-telegram-bot==6.1b2
vk
numpy